Distinguish between ordered and unordered card lists
Had a nice chat with Nirzar around card lists and the differing styles
for read more and search/nearby etc..

If order is important then the cards should be kept in a single list.

Random and Read more have no order so take the 3 card per row view.
Add a margin bottom for better separation of more than 3 cards.