Allow automatic edit summaries and edit warnings
When editSummary parameter is passed to EditOverlay it will no
longer prompt a user to provide one.

A editWarning parameter will allow the display of warnings in the
editing workflow - currently shown before save is clicked. We can use
this to check editors know what they are doing before hitting save.

Additional changes:
* If wikitext parameter is provided and user hits save the wikitext
will be submitted as is